structured vehicle routes and controller
